--- lib/app/modules/help-avisos/repositories/help_avisos_repository_interface.dart
@@ -1,5 +1,0 @@
-import 'package:system_maua_front/app/modules/help-avisos/models/avisos_model.dart';
-
-abstract class IHelpAvisosRepository {
-  Future<List<AvisosModel>> getInformacaoAvisos();
-}
