+++ lib/app/modules/help/pages/help-avisos/help_avisos_controller.dart
@@ -1,7 +1,8 @@
 import 'package:mobx/mobx.dart';
-import 'package:system_maua_front/app/modules/help-avisos/models/avisos_panel_model.dart';
-import 'package:system_maua_front/app/modules/help-avisos/models/avisos_model.dart';
-import 'package:system_maua_front/app/modules/help-avisos/repositories/help_avisos_repository_interface.dart';
+import 'package:system_maua_front/app/modules/help/pages/help-avisos/repositories/help_avisos_repository_interface.dart';
+
+import 'models/avisos_model.dart';
+import 'models/avisos_panel_model.dart';
 
 part 'help_avisos_controller.g.dart';
