+++ lib/app/modules/help/pages/help-avisos/help_avisos_page.dart
@@ -1,15 +1,16 @@
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:flutter_modular/flutter_modular.dart';
-import 'package:system_maua_front/app/modules/help-avisos/help_avisos_controller.dart';
+
 import 'package:flutter/material.dart';
-import 'package:system_maua_front/app/modules/help-avisos/widgets/avisos_widget.dart';
+import 'package:system_maua_front/app/modules/help/pages/help-avisos/widgets/avisos_widget.dart';
+
 import 'package:system_maua_front/app/shared/components/app_bar/app_bar_widget.dart';
 import 'package:system_maua_front/app/shared/themes/app_text_styles.dart';
 
+import 'help_avisos_controller.dart';
+
 class HelpAvisosPage extends StatefulWidget {
-  final String title;
-  const HelpAvisosPage({Key? key, this.title = 'HelpAvisosPage'})
-      : super(key: key);
+  const HelpAvisosPage({Key? key}) : super(key: key);
   @override
   HelpAvisosPageState createState() => HelpAvisosPageState();
 }
