+++ lib/app/modules/help/pages/help-avisos/repositories/help_avisos_repository.dart
@@ -1,5 +1,6 @@
-import 'package:system_maua_front/app/modules/help-avisos/models/avisos_model.dart';
-import 'package:system_maua_front/app/modules/help-avisos/repositories/help_avisos_repository_interface.dart';
+import 'package:system_maua_front/app/modules/help/pages/help-avisos/models/avisos_model.dart';
+
+import 'help_avisos_repository_interface.dart';
 
 class HelpAvisosRepository extends IHelpAvisosRepository {
   List<AvisosModel> listAvisos = [
