+++ lib/app/modules/help/widgets/help_card.dart
