+++ lib/app/modules/landing/landing_module.dart
@@ -1,10 +1,7 @@
 import 'package:flutter_modular/flutter_modular.dart';
-import 'package:system_maua_front/app/modules/help-aluno/help_aluno_module.dart';
-import 'package:system_maua_front/app/modules/help-avisos/help_avisos_module.dart';
 import 'package:system_maua_front/app/modules/activities/activities_module.dart';
 import 'package:system_maua_front/app/modules/calendario/calendario_module.dart';
-import 'package:system_maua_front/app/modules/help-bolsas/help_bolsas_module.dart';
-import 'package:system_maua_front/app/modules/help-intercambio/help_intercambio_module.dart';
+import 'package:system_maua_front/app/modules/help/help_module.dart';
 import 'package:system_maua_front/app/modules/home-aluno/home_aluno_module.dart';
 import 'package:system_maua_front/app/modules/landing/landing_page.dart';
 import 'package:system_maua_front/app/modules/materias/materias_module.dart';
@@ -22,14 +19,11 @@ class LandingModule extends Module {
         children: [
           ModuleRoute('/notas', module: NotasAlunoModule()),
           ModuleRoute('/home', module: HomeAlunoModule()),
-          ModuleRoute('/help', module: HelpAlunoModule()),
-          ModuleRoute('/avisos', module: HelpAvisosModule()),
+          ModuleRoute('/help', module: HelpModule()),
           ModuleRoute('/activities', module: ActivitiesModule()),
           ModuleRoute('/mural', module: MuralDivulgacaoModule()),
           ModuleRoute('/calendario', module: CalendarioModule()),
           ModuleRoute('/materias', module: MateriasModule()),
-          ModuleRoute('/bolsas', module: HelpBolsasModule()),
-          ModuleRoute('/intercambio', module: HelpIntercambioModule()),
         ]),
   ];
 }
