+++ test/app/modules/help/help-avisos/help_avisos_controller_test.mocks.dart
@@ -1,13 +1,13 @@
 // Mocks generated by Mockito 5.0.10 from annotations
-// in system_maua_front/test/app/modules/help-avisos/help_avisos_controller_test.dart.
+// in system_maua_front/test/app/modules/help/help-avisos/help_avisos_controller_test.dart.
 // Do not manually edit this file.
 
 import 'dart:async' as _i3;
 
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:system_maua_front/app/modules/help-avisos/models/avisos_model.dart'
+import 'package:system_maua_front/app/modules/help/pages/help-avisos/models/avisos_model.dart'
     as _i4;
-import 'package:system_maua_front/app/modules/help-avisos/repositories/help_avisos_repository_interface.dart'
+import 'package:system_maua_front/app/modules/help/pages/help-avisos/repositories/help_avisos_repository_interface.dart'
     as _i2;
 
 // ignore_for_file: avoid_redundant_argument_values
