+++ test/app/modules/help/help-avisos/help_avisos_repository_mock_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:system_maua_front/app/modules/help-avisos/repositories/help_avisos_repository.dart';
+import 'package:system_maua_front/app/modules/help/pages/help-avisos/repositories/help_avisos_repository.dart';
 
 void main() {
   late HelpAvisosRepository helpAvisosRepository;
